Make cd93 a reform
