+++ openfisca_france/model/model.py
@@ -64,10 +64,6 @@ from prestations.prestations_familiales import (  # noqa analysis:ignore
     cf,
     )
 
-from prestations.prestations_locales import (  # noqa analysis:ignore
-    cd93,
-    )
-
 from revenus import autres
 
 from revenus.activite import (  # noqa analysis:ignore
